fix(embroider): use factory function instead of dynamic importing
